remove migrations
