+++ runtime/litentry/src/lib.rs
@@ -134,7 +134,6 @@ pub type Executive = frame_executive::Executive<
 	// It was reverse order before.
 	// See the comment before collation related pallets too.
 	AllPalletsWithSystem,
-	(migration::FixParachainStakingStorage<Runtime>,),
 >;
 
 impl_opaque_keys! {
