+++ runtime/litentry/src/migration/migration.md
@@ -1,5 +1,10 @@
-# Migrate
-Under fiels P9191
+# Migrate fix of last decimal upgrade
+P9192.rs
+it fixes the missing `total` migration in TopDelegations
+However, we sorted the `CandidatePool` mistakenly by the backed stakings, which caused the same collator to be inserted twice in the vector. We've fixed that by forcefully setting the raw storage of `CandidatePool` via `system.SetStorage`.
+
+# Migrate decimal change 12 -> 18
+P9191/ folder:
 The migration including the following pallets:
 Minor pallet migration
 Bounty, Democracy, Identity, Multisig, Preimage, Proxy, Treasury, Vesting
